pull-request: prevent disappering reviewers when new commits added
Currently, completed reviewer list is reset when new commits added
to the contributor's pull-request branch. This commit will prevent
that situation with ongoing reviwer feature.
When a new commit added to the branch which used for pull-request,
completed reviewers moved to ongoing reviewer list (instead of reset)

TODO: When a completed reviewer moved to ongoing reviewer,
It is required to send notification.